+++ src/lib.rs
@@ -68,6 +68,7 @@ mod record;
 
 pub use crate::err::*;
 pub use model::*;
+
 use tokio::sync::mpsc::UnboundedSender;
 use tokio_stream::{wrappers::UnboundedReceiverStream, Stream};