+++ src/model/device_state.rs
@@ -12,7 +12,7 @@ pub enum DeviceState {
     TestRecord(TestRecordState),
     SolarCharger(SolarChargerState),
     BatteryMonitor(BatteryMonitorState),
-    InverterState(InverterState),
+    Inverter(InverterState),
 }
 
 impl DeviceState {
@@ -27,7 +27,7 @@ impl DeviceState {
             RECORD_TYPE_BATTERY_MONITOR => Ok(Self::BatteryMonitor(BatteryMonitorState::parse(
                 &record.decrypt()?,
             )?)),
-            RECORD_TYPE_INVERTER => Ok(Self::InverterState(InverterState::parse(
+            RECORD_TYPE_INVERTER => Ok(Self::Inverter(InverterState::parse(
                 &record.decrypt()?,
             )?)),
             _ => Err(Error::UnsupportedDeviceType(record.record_type())),