+++ src/model/mod.rs
@@ -14,3 +14,4 @@ pub use error_state::ErrorState;
 pub use inverter_state::InverterState;
 pub use mode::Mode;
 pub use solar_charger_state::SolarChargerState;
+pub use test_record_state::TestRecordState;